+++ packages/video_player_videohole/tizen/src/ecore_wl2_window_proxy.cc
@@ -37,21 +37,6 @@ void EcoreWl2WindowProxy::ecore_wl2_window_geometry_get(void *window, int *x,
   ecore_wl2_window_geometry_get(window, x, y, width, height);
 }
 
-int EcoreWl2WindowProxy::ecore_wl2_window_surface_id_get(void *window) {
-  if (!ecore_wl2_window_handle_) {
-    LOG_ERROR("ecore_wl2_window_handle_ not valid");
-    return -1;
-  }
-  FuncEcoreWl2WindowSurfaceIdGet ecore_wl2_window_surface_id_get =
-      reinterpret_cast<FuncEcoreWl2WindowSurfaceIdGet>(
-          dlsym(ecore_wl2_window_handle_, "ecore_wl2_window_surface_id_get"));
-  if (!ecore_wl2_window_surface_id_get) {
-    LOG_ERROR("Fail to find ecore_wl2_window_surface_id_get.");
-    return -1;
-  }
-  return ecore_wl2_window_surface_id_get(window);
-}
-
 EcoreWl2WindowProxy::~EcoreWl2WindowProxy() {
   if (ecore_wl2_window_handle_) {
     dlclose(ecore_wl2_window_handle_);